+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Update using ```composer update```.
 * ```Collection::filter()``` callables can now accept the item key as a second argument.
 * It is now possible to auto assign variables to views using the ```ViewFactory::autoAssign()``` method.
 * JSON responses can now set the response status code.
+* It is now possible to capture output in a template using the new `{% capture:name %}...{% endcapture %}` blocks.
 * Mako is now using Whoops for displaying exception details.
 
 #### Changes